[luciexe/build] Make `log` empty by default, and remove link if empty.
Currently luciexe's using this library will see a `log` link under EVERY
step, and this log for a SUCCESS-ful step will contain a single,
unhelpful line saying that the step was, in fact, SUCCESS-ful.

This change removes the unhelpful log line, and also adds functionality
to remove the "log" link from a step, if it turns out that nothing ended
up using it. The "log" link will only be removed at the conclusion of
the step, meaning that it will be present for builds which start a step and then crash. The alternate design (adding the link at the end, but
only if it's used) would have the potential to hide important log
information if the build crashes.

This change also will skip the `set status: SUCCESS` line if a step (or
the overall build) ends successfully.

The combination of these means that the vast majority of `log` links
will be gone from builds using `go/luciexe/build`.

R=azrahman@google.com, yiwzhang

Change-Id: Ia38bbaba22099ef63e4e0888f4ce7784f342e7e0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4199062
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>