[milo] mount build page to react router
Mount build page and related tabs to react router.

Switching between tabs will cause a full page reload.
The best way to solve this is to migrate the build page component to
React and use the <Link /> component from react-router for navigation.
This will be implemented in the following CL.

R=mdraz@google.com, nqmtuan@google.com

Bug: 1408955
Change-Id: Iaa2f16c00e88a6f6afdc8d1a55288ccc6a14385c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4278781
Reviewed-by: Mostafa Draz <mdraz@google.com>
Commit-Queue: Weiwei Lin <weiweilin@google.com>