[auth-service]: Update AuthRealmsGlobals singleton.
This one is kinda awkward, since the proto is Marshalled by the Python datastore side. By the time we get the proto on our side it's automatically in a partially unmarshalled format but not in the correct form we want. As you'll see I manually also unmarshal it. As a result, this means that I can't write back for now because I can't marshal it back to the format that it came in so the datastore entity is just logged. The reason for having the commitEntity there is for testing and for when I actually teach Python how to work with permissions.cfg then we will actually write over the old one and forget about it.

Bug:1336137
Change-Id: Ia14bf547fa33f2b56c551565d933eadccedfe221
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4285740
Commit-Queue: Camilo Jacomet <cjacomet@google.com>
Reviewed-by: Yulan Lin <yulanlin@google.com>
Reviewed-by: Dan Le Febvre <dlf@google.com>