[auth-service]: RealmsInternals, GetConfig.
This is the first of a few CL's that will deal with expanding the realms of projects into the internal representation.
The internal representation is how the realms are bundled into AuthDB when clients use AuthDB. This part just handles getting all
the respective realms.cfg files concurrently and returning them to the caller.

Change-Id: Ice6a9d4af0e482e1296a4efeea1927119ff9192c
Bug:1336137
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4336747
Reviewed-by: Yulan Lin <yulanlin@google.com>
Commit-Queue: Camilo Jacomet <cjacomet@google.com>
Reviewed-by: Dan Le Febvre <dlf@google.com>