[Analysis] Improve handling of unavailable rule definitions in the UI.
Some users are unable to see rule definitions, because they do not
have analysis.rules.getDefinition permission. Currently, this
results in blanks on the rules page and rule page. This makes it
impossible to navigate to the rule from the rules page.

After this change, we display some filler text instead of showing
nothing.

BUG=None
TEST=npm run test

Change-Id: I67b2b5dffaaafa72b0c87384c2237eccc967984e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4364403
Reviewed-by: Mostafa Draz <mdraz@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>