[encryptedcookies] Add a standard endpoint to serve tokens.
It is "GET /auth/openid/state" that returns JSON with session
state (including access and ID tokens).

It has been "implemented" at least 4 times already. It will be
needed for Swarming as well. Instead of adding the fifth
copy-pasta, move the implementation into the base module.

It is based on the code in
go.chromium.org/luci/analysis/frontend/handlers/auth_state.go
which seems to be the original source of most copy-pasted
snippets.

It should be wire-compatible with all these copies. The only
change is that it adds `accessTokenExpiresIn` and
`idTokenExpiresIn` which are useful during debugging to
eye-ball when tokens expire (especially in presence of skewed
clocks).

This endpoint is disabled by default, since it makes XSSs scarier.
It can be enabled by passing easily memorable and succinct flag:
  -encrypted-cookies-expose-state-endpoint

R=meiring@google.com, weiweilin@google.com

Change-Id: I441f5409ef6ce6c53db0f10fe1d748ddb4e89441
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4402093
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>