[logdog] Remove two indexes for Created field.
This removes the built-in index for Created and also the unused query
functionality for this field.

Previously the query API would sort by created time, in order to expose
the ability to search for logs created in a certain time range. This
functionality is much less useful than it sounds, though, because this
was constrained to log streams within a single prefix.

Should someone need this, it would be feasible to just scan all the
log streams within the prefix to see their creation times.

Note that this will change the order of logdog queries from Created
time order to ordered by a hash of the prefix/stream_name. However,
the previous order was arbitrary and undocumented, the new order is
also arbitrary and undocumented.

I'm making this change to assist the Datastore->Firestore migration,
because currently the Created index is quite hot; Newly created streams
happen constantly, which means this index is likely getting updated at
a high rate, and fairly linearly.

R=ddoman, mwarton, vadimsh

Change-Id: Iadd748bfe95120dd35f5517563abeb797c7faf99
Bug: 1430940
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4401279
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>