[ResultDB] Update (Query+BatchGet)TestVariants to expose sources.
The endpoints now report which code sources were tested.

Originally, the plan was to update QueryTestVariants only, but
this turned out to be harder than updating both (given the shared
use of testvariants.Query and the Test Variant proto), so we are
updating both.

BUG=b:274041961
TEST=Integration tests

Change-Id: Ie7470ebda1264f20c089aaeda008f57c98172168
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4401783
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>