[lucicfg] Rationalize GetRewriter inputs to be absolute.
Previously GetRewriter accepted `path` which could be:
  1) Relative to the user's $CWD
  2) Relative to the entrypoint (e.g. main.star) file directory
  3) Absolute.

Cases 1 and 3 could happen for `lucicfg fmt` and `lucicfg lint`, e.g.

   lucicfg fmt ../somefile.star
   lucicfg fmt /path/to/somefile.star

Case 2 happened for `lucicfg validate` whether or not the argument
on the CLI was relative or not, e.g. these were the same, because
the Validate process was asking for formatters for paths derived
from the load statements:

   lucicfg validate ../main.star
   lucicfg validate /path/to/main.star
   (e.g. a statement in main.star like `load("//some.star", ...)`
    would yield a path "some.star")

Unfortunately, there was a bug where `lucicfg fmt ../somefile.star`
would attempt (and fail) to resolve the common ancestor for
`/path/to/../somefile.star` if there was a .lucicfgfmtrc file.
This was because GetRewriter was adjusting all
relative paths to be relative to the config file regardless of
context.

This CL changes GetRewriter to only accept absolute paths, and also
changes fmt, validate and lint to only provide absolute paths, thus
resolving the ambiguity in all cases.

We considered adding additional state to RewriterFactory to allow
GetRewriter to absolutize paths differently in different context
but decided against it to reduce the combinatorics of the
implementation.

Bug: 1417086
Change-Id: Ic0b13ef9e96922bc2684aba17183ff2e54401ed1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4454752
Reviewed-by: Riley Wong <rgw@google.com>
Commit-Queue: Riley Wong <rgw@google.com>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>