gce: Start the swarming bot service after cloud-init
In https://crrev.com/i/5808689, the chromebuild-setup service on GCE
instances will correct the OS's settings to properly categorize an
SSD boot disk as non-rotational. Swarming uses this setting to fill
the 'ssd' dimension on bots, and permanently memoizes that val.

But given the current ordering of services in systemd, the swarming bot
will start a good 5 seconds before the chromebuild-setup service.
This leads to swarming reading the system settings before
chromebuild-setup has a chance to correct them, causing the swarming
bot to be permanently confused.

The reason swarming starts so much sooner is that chromebuild-setup
is initiated by cloud-init.target, while swarming is started by
multi-user.target. The former seems to occur at a checkpoint later
on in the instance's boot-up. So this changes swarming's service to
match the "WantedBy" and "After" fields of chromebuild-setup. This
should be safe for swarming since the network should already be up
by then, but it will make GCE bots connect to swarming ~5s later than
they currently do.

That alone would make the two services start at the same time, but
we also want the swarming bot service to start _after_ the
chromebuild-setup service. That can be partly achieved by either
putting a "Before=swarming-start-bot" directive in chromebuild-setup,
or by putting a "After=chromebuild-setup" in the swarming service.
We'll go with the former to keep swarming's config agnostic of a
chrome-specific service. In other words, it's safe to assume that
swarming-start-bot will run everywhere that chromebuild-setup runs.
But the reverse might not be safe.

And then a final piece needed to get the swarming bot to start
after chromebuild-setup has finished is to change chromebuild-setup
to a "oneshot" service (sourced from
https://trstringer.com/simple-vs-oneshot-systemd-service/). Since
chromebuild-setup's service cfg file is in another
repo, that will be done elsewhere.

Bug: b/279078023
Change-Id: I555e0b60d9425d7cbbb41373d7b514847e24ab3a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4492133
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>