Extend location tags for file loc
dirmd is updated to support a per-file override definition, which
are mapped to Files map[string]metadata.

If the test result file loc matches an entry in the mapping.Files
map, this would indicate that the metadata has been overridden for
that specific file and should use that metadata.

Bug: 1431748
Change-Id: Ic1cb2fc36d78acb8d909ec6111ee401ae3f28933
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4472948
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>