[Analysis] Use int64 instead of uint64 for GitilesCommit.position.
uint64 is not compatible with the BigQuery Integer type. The
additional range afforded by this type over int64 is not really
important, so drop it back to an int64 and use validation to
prevent use of negative values.

Proto is quite new and currently only used in pRPC transfers, so
replacement should be safe.

BUG=b:279110567
TEST=ResultDB + LUCI Analysis Integration tests.

Change-Id: I72f02828c13f2505c7de3bf642469bf587f04e2f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4501680
Commit-Queue: Beining Chen <beining@google.com>
Auto-Submit: Patrick Meiring <meiring@google.com>
Reviewed-by: Beining Chen <beining@google.com>