[Analysis] Update BigQuery batching to respect size limits.
BigQuery has request size limit of 10 MB on Storage Write API calls.
We currently exceed this in some cases with the verdict export,
because our batches are based on a fixed number of rows, not a
fixed maximum size.

Update batching to generate batches based on a maximum 9 MB of
rows each time.

BUG=b:279110567
TEST=Integration tests

Change-Id: I1d30d4240f275499ee8dcaa8e55fa6c4bf458f6d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4505271
Reviewed-by: Beining Chen <beining@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>