[cipd] Bump up maximum number of cache entries to 10k.
felizardo@ discovered a problem in some of their builders where it seems
that CIPD is running into the 'rolling eviction' problem - the task is
attempting to `cipd ensure` ~530 packages, and seems to always be
downloading all of them, so we suspect that this is the culprit.

Note that the cache eviction also defaults to a maximum of 2 days for
unused entries. We suspect that in practice, this will be enough to keep
the cache in decent shape, even with the higher total number of entries.

Followup to this would be to change the cache to be size-based instead,
and allow the maximum disk usage to be set on a per-machine level
(likely using a function on the current disk size as a default).

R=bryner, vadimsh, felizardo

Change-Id: I13f3697d56ab1acd2a1442ba02195672f0e3f167
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4523075
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>