milo: test history links use kv pairs instead of vhash
vhash is meant to be an internal only concept.  We shouldn't expose this
through links into the filter box in the UI.

There is a small chance that the substring matching used will show more
than a single variant, but this shouldn't be a big usability issue even
if it happens.

On the other hand, having the key value pairs allows the user to
edit/remove them to see and compare to other interesting variants.

Bug: b:241916558
Change-Id: I16e58e4788553362048de7a03d083ab7b411fd42
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4535444
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Auto-Submit: Matthew Warton <mwarton@google.com>
Commit-Queue: Weiwei Lin <weiweilin@google.com>