[swarming] Ignore expired poll token if there's a session token.
If a Swarming bot executes a task for >1h (enough to expire the
poll token), and then gets a `terminate` or `update` command from
Swarming, it attempts to close RBE session sending expired poll
token (because /bots/poll doesn't produce a new one, it does so
only when replying with `rbe` command).

One approach is to change the Swarming backend to always return
a poll token, with every command. But this will make bot code
somewhat more complicated.

Instead we'll just fallback on a session token (which is kept
fresh by `/bot/rbe/session/update` even when the task is running).
The poll state in it might be out-of-date, but it should not
matter if the bot is closing the session anyway.

There's no security implications, since if the caller has
a valid session token and expired poll token, they can just
avoid sending the poll token and get the same result as when
sending both tokens (after this CL lands).

Also slightly adjust logging to simplify debugging token staleness.

R=jonahhooper@google.com
BUG=1377118

Change-Id: I39a6a2c519b9c37e86a710468906fac22a39d1e2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4546237
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>