[Analysis] Refactor TestVariantBranches API, hide Spanner proto.
Remove the ability of clients to depend on the internal Spanner
storage proto. The fields disclosing it now use anypb to avoid
needing to explicitly reference the storage proto, and comments
indicate the output is for debugging purposes only.

If actual clients need to consume it, we should define a proper
wire proto and use that.

See also go/apidosdonts.

BUG=b:284771604
TEST=Integration tests

Change-Id: I9453c224d3082d5eeb36706112d40b3fe78798c8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4569669
Reviewed-by: Tuan Nguyen <nqmtuan@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>