cv: use reset trigger long op during submission instead directly reset in place
This is to converge the code path so that we can easily emit metrics for
resetting the trigger and properly handle the reset trigger failure

It also has another behavior change that the error for posting dependent
CL submission failure on submitted CL will be swallowed because those
messages are not critical. Later on, I will move those to long op with
better mechanism.

R=ddoman

Bug: 1417328
Change-Id: I56821af15fe2f3605f10e1d01223b658459bab2c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4576329
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>