config: rename a few things
1. rename ValidateConfig rpc to ValidateConfigs rpc because LUCI
Config will pass multiple configuration files to the applications
2. rename ValidateConfigResponse message to ValidationResult
because it will be used as the type of a field in the ValidateConfigs
rpc on LUCI Config service as well (see the subsequent CL). A response
message in another response message may seems confusing.

Bug: 1446032
Change-Id: Idae4c314a6c6aaa3e20c829577ea38c911623958
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4574016
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>