[cv] remove permanent_errs and posted from eventpb.PostStartMessage
The fields are currently used for test verification purposes only.
- No logs use the fields.
- No RunLogs print the field values.
- No logic alternating the behaviour, except that PostStartMessageOp.DoCL()
uses the field values to determine the op status. However, the op status
can be determined without the fields in the task proto.

However, even in the unit tests, the fields aren't necessary to validate
the outcome of PostStartMessageOp.Do(). Checking the last message posted
on gerritfake should be enough.

Change-Id: I5ec42d23cb85d90d2593bd16a5e265dda25de566
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4585809
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>