[cv] resurrect Convey() for TestPostStartMessage:failures
This CL simplifies the tests with testing only one single GerritCL
for a given Gerrit error.

(1) PostStart doesn't perform the Gerrit post operation directly.
It's handled within the gerrit module (or gerritfake), and it should
be safe to assume that no messages will be posted as a side-effect of
the ci.ACL override.
i.e., if ci.ACL returns an error, no messages should be posted.

This CL still checks the number of messages posted, but working with
one single CL should be enough to check the validity of the side effect.
I don't see how testing with a mix of multiple CLs in TestPostStartMessage
improves the test coverage and quality better.

(2) PostStart no longer returns different values for transient and non-tr
errors.
The tests were reasonable because PostStart() used to return different
status code, depending on whether the Gerrit error was transient or not.
The logic was removed, and it's no longer applicable to test the code path
with a mix of successful, non-transient, and transient within the same
Convey(). In other words, testing each error type in a separate Convey()
is enough.

(3) There is no completely safe way to enforce the sequence of the Gerrit
operation between CL1 and CL2 in the current testing interface.

Channels can be used inside c.ACLs(), but it cannot strongly enforce
the order of Gerrit op completion between CL1 and CL2. It can lower
the chance much by letting c.ACLs on CL2 wait until c.ACLs() on CL1 with
op == gerritfake.OpReview is returned. However, it is fragile because
it relies on the implementation of doCL() too much and a test failure
wouldn't be super intuitive, as it would just give timeout in the test
output.

In short, it's not worth complicating the tests with mocking and testing
multiple changes within the same Convey().

Bug: 1453166
Change-Id: Ib320811e304dbe1297a50d2097183643b58f224d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4605995
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Scott Lee <ddoman@chromium.org>
Auto-Submit: Scott Lee <ddoman@chromium.org>