[common/exec] Add a mockable drop-in replacement for "os/exec".
The intent of this is to allow really easy mocking for very heavily
subprocess-driven programs (i.e. "recipes").

This library works by pattern-matching the to-be-executed *Cmd with
mocks in context, gob-encoding them over http to a re-invocation of
the current executable (i.e. test binary) which is set to intercept
such invocations in TestMain().

In production this passes through all functionality to "os/exec".

Change-Id: Ib1ab4b1b148cd821d0453fec2c2b47e8215bd961
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3814638
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>