[Analysis] Minor UI refinements.
Change the "Why?" link on the exonerations tab to a button for
consistency with the recommended priority tab.

Fix grammar around the rule audit info if only the created/
last modified time is known but not the user.

BUG=None
TEST=npm run test

Change-Id: Ic551515e4b7795aa6ab531eb8e063cda4dad6e3f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4626136
Reviewed-by: Matthew Warton <mwarton@google.com>
Auto-Submit: Patrick Meiring <meiring@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>