[gce-provider] Add a heuristic to determine leaks for deletion
Bug: b/274688233

This patch adds a heuristic approach to determining if the VM was leaked
by gce-provider. The approach is to determine if there exists a record
in gce-provider that could have created the VM. If there is we can
drain the bot safely. The downsides for this approach is that, if the
record doesn't exist for some reason, say the pool was resized, renamed
or removed. Then those VMs are left alone to exist.

Change-Id: I40913997cd5e20feaa3fc533d61ee09f173d8bca
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4628140
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>