[buildbucket] having UpdateBuildTask always update Task Status, UpdateID, and SummaryHTML, while optionally updating other fields only if provided.
Previously, when a task was sent via the BuildTaskUpdate message in pubsub, then the entire task 
would be stored in datastore.That would overwrite things that came from RunTask (like the task's 
link). To make BuildTaskUpdate more lightweight (and backends can focus on sending only a status if desired), we 
need to ensure that the link and other details are not unintentionally overwritten by UpdateBuildTask.

Bug:1430448
Change-Id: Ia1a3d7485c28a21727a10c2b2e39645bfaad4da7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4632162
Commit-Queue: Randy Maldonado <randymaldonado@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>