[Buildbucket] StartBuild for build on swarming: support no task id on file
It's possible that swarming successfully creates a task, but buildbucket fails to save it in datastore.
In this case, a StartBuild request could reach to Buildbucket before the retry of CreateSwarmingBuildTask, with no task id saved in datastore.
And StartBuild should link the swarming task with the build.

Also in this CL: swarming task creation is idempotent now, so no need to cancel the already created swarming task and create a new one if datastore save fails.

Bug: 1455222
Change-Id: I6709b1d3385c406d557bfc7c5029a2e3ca66c53a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4642163
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Commit-Queue: Chan Li <chanli@chromium.org>