Add baseline as a data field to invocations
A baseline is a user defined string. It represents a set of test
variants which should be part of the submitted code for a project.
It's expected that each CQ try builder will have its own baseline
for Chrome.

This change introduces the Baseline spanner table, which will be
used to store Baselines. Baselines within 72 hours of the
LastUpdated field will be considered in "spin up state" and are
intended to be used as part of future query calls when out of this
band of time.

This change also implements baselines as part of Invocations.
Create and update invocation calls will be able to add baselines.
Validation is introduced as part of this change, defined by the
regex in pbutil. Baseline should be specifiable through both
the Recorder RPC and CLI.

Bug: 1456545
Change-Id: Iff3f34ce9b3e2ebb29134f559a95079af730a7d1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4628865
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Auto-Submit: Jeff Yoon <jeffyoon@google.com>