[auth-service]: RealmsExpansion basic operations.
This is the first CL for RealmsExpansion, we already have in place the Roles Expansion and ConditionsSet expansion. This will be used by config to expand the Realms to the internal representation. Tests include testing for errors, testing of results will be in config_test.go in follow up.

Bug:1336137
Change-Id: I6ccf3ae985bd7d6d1685bf93e4b16b82aa697029
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4519300
Reviewed-by: Joey Scarr <jsca@google.com>
Commit-Queue: Camilo Jacomet <cjacomet@google.com>