[Analysis] Fix corruption in changepoint analysis export.
References to `tvb` were escaping testvariantbranch.ReadF,
which is contrary to the contract of ReadF, and causes
problems as the buffer is re-used. This caused segment data
to be exported to BigQuery against the wrong test variant
branch.

To avoid a recurrence, end to end changepoint analysis
ingestion tests have been improved to verify BigQuery export
is correct when multiple test variants are ingested.

BUG=None
TEST=Integration tests

Change-Id: I1c5c13805bc1c74e3ef608404d1aa62012fecad4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4656173
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Tuan Nguyen <nqmtuan@google.com>