[milo] add self.createSelectiveSpiesFromModule
This is similar to self.createSelectiveMocksFromModule except that
1. the original implementation is used by default, and
2. as a result, only functions can be spied (class methods can easily be
spied using `jest.spyOn`).

This is useful when we want to check whether a top-level function from
a module is called without changing the behavior of that function (e.g.
check whether `usePrpcQuery` is called with the correct params).

R=mdraz@google.com, nqmtuan@google.com

Bug: None
Change-Id: I043b87cbf7bd46ccd34f00eeaaa0f6ed5e385ac3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4659090
Reviewed-by: Mostafa Draz <mdraz@google.com>
Auto-Submit: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Mostafa Draz <mdraz@google.com>