ResultDB MarkInvocationSubmitted RPC
This change introduces the RPC to be able to mark invocations as
submitted. Marking an invocation submitted is adds all test variants
from that invocation to the set of tests for its baseline. The set
of tests for that baseline are tracked in the BaselineTestVariants
Spanner table.

Marking an invocation submitted is asynchronous and utilizes the
task queue to accomplish this. All invocations must be finalized
prior to it being marked submitted.

Change-Id: I0d444ffd9c8bfe8dc72338d06718a50d1703f735
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4644092
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>