[ResultDB] Add submitted status for invocations
Invocations that are marked submitted should only be processed
when it is finalized. When invocations are marked submitted, but
the invocation is not finalized, the mark submitted RPC is updated
to simply mark the invocation "submitted" and return.

In this case, it's expected that the finalizer will schedule the task
to ensure that the test variants associated with that invocation that
has been marked submitted to be added.

Submitted values are not backfilled, so a nil submitted status will
resolve to false.

Bug: 1456545
Change-Id: Id436b3098fcd4a6508ea8136987f168db5659866
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4674030
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>