[led] rename edit-cr-cl to edit-gerrit-cl and update docs
Right now the help output and name of edit-cr-cl strongly implies that
it's somehow Chrome-specific, and although it once was it isn't anymore.
Move the edit-cr-cl command behind the "advanced" options, add a
deprecation notice, and add the edit-gerrit-cl command which does
exactly the same thing, just with a nicer name and docs reflecting the
current state of things.

Also, add a copyright header to the tests. I can't pass presubmit
otherwise.

Bug: 1463082
Change-Id: I12543eb12d90a18ed6678bd659f99bf0010284a9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4673177
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Michael Knyszek <mknyszek@google.com>