[Analysis] Show a better error if the user cannot view rule for bug.
If the user cannot see the rule for a bug, provide a better message.
Currently, a user who is not logged in is told the rule does not
exist when they simply do not have permission to view it.

Under the new behaviour:
- Users who are not logged in, are always prompted to log-in.
- If the user is logged in, both the possibility the rule does
  not exist and possiblity they do not have permission see the rule
  is mentioned.

BUG=b:290702386
TEST=None

Change-Id: I8b7c7bb4abee7fc62181c349ce87fe6ddc9e2e66
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4675137
Reviewed-by: Mostafa Draz <mdraz@google.com>
Auto-Submit: Patrick Meiring <meiring@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>