resultdb: move the scalability check further down the code paths
QueryTestVariants is hitting the scalability limit of 20,000 invocations
on some queries.  However, most of these invcoations fo not have test
results.

Therefore, if we push the 20,000 check to be the ones that
QueryTestVariants is actually going to pass to Spanner (ie the ones with
test results), we will allow the calls to complete successfully.

Bug: b:291444564
Change-Id: Id7234078f82a9c3a7e656efcd97c3bc6c49d1488
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4686017
Auto-Submit: Matthew Warton <mwarton@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>