[CV] Track ResultDB metadata with tryjobs
BuildInfra from Buildbucket contains infra specific metadata.
For this change, the metadata of interest is ResultDB hostname
and invocation, which are specifically added to the list of
fields to parse.

This change introduces an infra field to a tryjob's result
buildbucket metadata. Although the definition follows buildinfra,
we restrict it only to resultdb.

This will allow CV to know which ResultDB to call with which
invocation when calling ResultDB's MarkInvocationSubmitted.

Bug: 1456545
Change-Id: I1c497f08c369441df23f17636c40f835507412eb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4696753
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>