[rpcexplorer] JSON parsing for the auto-completer.
Given a potentially incomplete JSON document and a position in it
we need to figure out what Protobuf object (per protobuf schema) is
being typed right now to populate the auto-completer list with its
fields. Additionally, if the field is partially typed already, we
need to recognized it and offer completing it.

E.g. parsing `{ "a": {"b": "some_fie` should tell us that we are
working with protobuf message matching type of <request>.a.b proto
field and we are interested in fields starting with `some_fie`
prefix.

A similar thing for auto-completing values (useful for enum-valued
fields).

I looked at ~4 existing JSON parsers (including streaming ones) and
they all don't handle this very well. In particular they don't like
we feed them incomplete JSON. Or they don't return enough
information to parse incomplete strings like `"some_fie` in the
example above.

I had to write my own. It is focused exclusively on this task.
It is not a general-purpose JSON parser.

R=randymaldonado@google.com

Change-Id: I25f2fa5d303a1643cec6b790ea6a39b2d7664129
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4642153
Reviewed-by: Jonah Hooper <jonahhooper@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>