[ResultDB] QueryNewTestVariants RPC
This change introduces the QueryNewTestVariants RPC to ResultDB, which
is used to calculate new tests.

Callers are expected to have resultdb.baselines.get to use this RPC.

Test variants become associated with a baseline through CV when it marks
invocations submitted after a change is successfully merged. This RPC
calculates the difference between test variants for a baseline
identifier and test variants for the invocation. The difference is returned
as "new tests".

If a baseline was created within 72 hours in the Baseline table, it's
considered to be in a "getting ready" stage. During this time, new tests
cannot be calculated. A baseline that's not present in the Baseline table
also indicates that the baseline has not been run for a while and has been
ejected from the table, so new tests should not be calculated.

Bug: 1456545
Change-Id: I4bf4009ef92c853b6b9cca07505a0654eba9ca31
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4719266
Reviewed-by: Patrick Meiring <meiring@google.com>
Auto-Submit: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>