[lucicfg] Expose a descriptor set with API annotation protos.
This is necessary in case a lucicfg script wants to work
with protos that reference e.g. google/api/field_behavior.proto.

It is not allowed to have multiple *FileDescriptorProto with
the same filename in a proto registry. To simplify dealing with
that lucicfg has a mechanism that allows to "share" a bunch of
descriptors as a single library-like unit called "descriptor set".

Such descriptor sets can have other descriptor sets as dependencies.
lucicfg will load any individual descriptor set only once.

This in particular is important for protos that are referenced by
lucicfg own stdlib (like LUCI protos and their dependencies). Their
descriptors are loaded into the default registry already. It means
that a custom descriptor set must not include them directly as
*FileDescriptorProto, but instead should refer to them via the
a descriptor set dependency (to essentially "deduplicate" loading
them).

This was all already in place for well-known protos like e.g.
"google/protobuf/descriptor.proto", but few other protos sneaked
in into the default registry as dependencies of LUCI protos,
in particular:
  * google/api/field_behavior.proto (from google.golang.org/genproto)
  * validate/validate.proto (from protoc-gen-validate)

They ended up in the descriptor set with LUCI protos. It means that
if a lucicfg script wants to use some custom protos that refer to
field_behavior.proto or validate.proto, they need to specify
`lucitypes_descpb` descriptor set as a dependency. This is not great
because:
  * It is unexpected and hard to discover.
  * Doing that will also import all the LUCI protos as well, which
    might be undesirable in some circumstances (when using a separate
    non-default proto loader that doesn't have all LUCI protos
    already).

This CL moves google/api/*.proto and validate/validate.proto into
two new descriptor sets, so they can be targeted individually and
also adds a check to prevent unexpected dependencies to sneak in
again.

R=yiwzhang@google.com
CC=​justinsuen@google.com

Change-Id: I7fcb36d2eb64feca6bc911ac7c5bb6a35495f54d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4741512
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>