[logdog] Stop cookie-based auth on `services` service.
Ensure cookie-based authentication is never used for this service.
The default GAEv1 cookies implementation stores email addresses
indefinitely. There are privacy concerns with this.

It shouldn't be used anyway, as this service should only be
used from the logdog backend. But just to be sure, we disable
it here.

BUG=None

Change-Id: I446dfc923372be2175a58dd06c1ac520911aaa60
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4749985
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Patrick Meiring <meiring@google.com>